requirements.txt
python==3.10
selenium==4.7.2
webdriver-manager==3.8.5
PyYAML==6.0
python-telegram-bot==13.15